requirements.txt
torch~=2.0.0
torchvision~=0.15.1
lightning~=2.1.2
wandb~=0.17.0
hydra-core~=1.3.2
omegaconf~=2.3.0
torchmetrics[image]==0.11.4
wandb-osh==1.2.1
gluonts[torch]==0.13.1
pytorchvideo~=0.1.5
colorama
tqdm
opencv-python
matplotlib
click
moviepy
imageio
einops
pandas
pyzmq
pyrealsense2
internetarchive

